requirements.txt
python-telegram-bot==21.7
gTTS==2.3.2
pyttsx3==2.90
colorama==0.4.6
python-dotenv==1.0.0
psutil==5.9.6